[YANG] remove uses clause handling, now part of sonic-yang-mgmt (#3814)
What I did
sonic-yang-mgmt now integrates uses clauses as if a uses clause didn't exist. No need to have this complex handling in sonic-utilities.

Depends on sonic-net/sonic-buildimage#21907
Fixes sonic-net/sonic-buildimage#22382

How I did it
Besides removal of the code, the only test case changes are the removal of the 'grouping' clause as it is specific to 'uses' which is not relevant and some minor reordering of expected responses.

How to verify it
Run tests/cli_autogen_yang_parser_test.py after applying both this PR and sonic-net/sonic-buildimage#21907