Secureboot: Image signing verification enhancements (#3989)
#### What I did

The current signature verification of sonic images assumes the DB Keys are all Root CAs.  The secureboot standard says nothing about this, the DBKeys are explicitly trusted by signing them with the KEK, and that signing method does not follow the standard X.509 PKI architecture.  Therefore the DB Key is not guaranteed to be a CA Root (aka not self-signed).  It is possible the DB Key was created as an intermediate, but since it is explicitly trusted that is ok.

Fixes https://github.com/sonic-net/sonic-buildimage/issues/23406

#### How I did it

This adds this explicit trust of the DB Key as a secondary signing verification if the original verification fails.  It disables looking inside the pkcs7 container for any keys at all and assumes the key specified is the exact key for the signature.

#### How to verify it

Build secureboot image signed with a DB Key that is not self-signed, then run through the sonic-installer install with that image and see the verification succeeds.


Signed-off-by: Brad House <bhouse@nexthop.ai>
Signed-off-by: Travis Brown <travisb@nexthop.ai>